Clarify board width/height
The example JSON can cause confusion since the JSON describe a game board with width/height of 11, but it's right under an example image of a board with a width/height of 12. This could lead developers to believe that the API returns a width/height that's off-by-one than it actually is (Totally didn't just happen to me :P). 

This change updates the example JSON to match what's being shown in the board image.