Add a note to 'next move' section.
When I was planning my logic the first time I assumed that the moves (left, right, up, down) were with respect to the direction of the head of the snake given the snake's head does move to point in different directions sometimes. Maybe adding a little note to the docs could answer this question earlier for some new players, it confused me more than it should have.